// internal/vibe-tests/src/types.ts
export const TARGET_NAMES = [
  'astryx',
  'astryx-tailwind',
  'astryx-svelte',
  'baseline',
  'html',
] as const;

export type TargetName = (typeof TARGET_NAMES)[number];